Create mnist_dataset_compress.py
A file used to load the MNIST dataset, this loads three arrays: training_data, validation_data and test_data. 

These three arrays then undergo a compressive sensing operation that employs a 2-D DCT bias function to transform
the images to the frequency domain. A random Gaussian matrix is used as the sensing matrix to select values from
the images which were transformed to the frequency domain. This compressed version of the vector is then passed to
our neural network for recognition